+++ R/req-perform.R
@@ -107,6 +107,7 @@ req_perform <- function(
 
   delay <- 0
   while (tries < max_tries && Sys.time() < deadline) {
+    retry_check_breaker(req, tries, error_call = error_call)
     sys_sleep(delay, "for retry backoff")
     n <- n + 1
 
@@ -161,6 +162,7 @@ handle_resp <- function(req, resp, error_call = caller_env()) {
 req_perform1 <- function(req, path = NULL, handle = NULL) {
   the$last_request <- req
   the$last_response <- NULL
+  signal(class = "httr2_perform")
 
   if (!is.null(path)) {
     res <- curl::curl_fetch_disk(req$url, path, handle)
@@ -228,6 +230,10 @@ last_request <- function() {
 #' actually sending anything. It requires the httpuv package because it
 #' works by sending the real HTTP request to a local webserver, thanks to
 #' the magic of [curl::curl_echo()].
+#'
+#' ## Limitations
+#'
+#' * The `Host` header is not respected.
 #'
 #' @inheritParams req_verbose
 #' @param quiet If `TRUE` doesn't print anything.