+++ tests/testthat/test-curl.R
@@ -51,7 +51,11 @@ test_that("can handle line breaks", {
 test_that("headers are parsed", {
   expect_equal(
     curl_normalize("curl http://x.com -H 'A: 1'")$headers,
-    as_headers("A: 1")
+    new_headers(list(A = "1"))
+  )
+  expect_equal(
+    curl_normalize("curl http://x.com -H 'B:'")$headers,
+    new_headers(list(B = ""))
   )
 })
 
@@ -136,6 +140,31 @@ test_that("can translate data", {
     curl_translate("curl http://example.com --data abcdef")
     curl_translate("curl http://example.com --data abcdef -H Content-Type:text/plain")
   })
+})
+
+test_that("can translate ocokies", {
+  skip_if(getRversion() < "4.1")
+
+  expect_snapshot({
+    curl_translate("curl 'http://test' -H 'Cookie: x=1; y=2;z=3'")
+  })
+})
+
+test_that("can translate json", {
+  skip_if(getRversion() < "4.1")
+
+  expect_snapshot({
+    curl_translate(r"--{curl http://example.com --data-raw '{"a": 1, "b": "text"}' -H Content-Type:application/json}--")
+    curl_translate(r"--{curl http://example.com --json '{"a": 1, "b": "text"}'}--")
+  })
+})
+
+test_that("content type stays in header if no data", {
+  skip_if(getRversion() < "4.1")
+
+  expect_snapshot(
+    curl_translate("curl http://example.com -H Content-Type:text/plain")
+  )
 })
 
 test_that("can evaluate simple calls", {